feat(openapi): Added openapi snapshot tests
